requirements.txt
Django==3.2
django-cors-headers==3.7.0
mssql-django==1.3
djangorestframework==3.13.0
djangorestframework-simplejwt==4.8.0
pyodbc==4.0.39
Pillow
django-filter==22.1
ldap3
xhtml2pdf
cx_Oracle
requests
PyMySQL==1.0.2
python-decouple
reportlab==4.2.5
openpyxl==3.0.10
msal==1.27.0
django-microsoft-auth==3.0.1